feat: use dynamic named ranges (fixes #7)
To avoid showing blanks in the drop-downs
(e.g. accounts, categories,and income sources)
the named ranges have been updated to use a formula
with OFFSET and COUNTA that sets the length to
exactly the number needed to fit all the non-blank entries.
See https://support.microsoft.com/en-us/help/830287/how-to-create-a-dynamic-defined-range-in-an-excel-worksheet

Note that this only works right if the entries are contiguous
(e.g. no gaps).